requirements.txt
datasets==2.12.0
pytorch-lightning==2.0.2
lightning==2.0.2
lightning-utilities==0.8.0
torch==2.0.1
torchmetrics==0.11.4
transformers==4.29.2
wandb==0.15.3
python-dotenv==1.0.0
numpy==1.24.3
matplotlib==3.7.1

